binance/requirements.txt
requests
tqdm
pandas
coinbase-advanced-py
PyYAML
